% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFFT.R
\name{buildFFT}
\alias{buildFFT}
\title{Title}
\usage{
buildFFT(.x = NULL, s = NULL, t = NULL, zp = 256, kf = 1)
}
\arguments{
\item{.x}{data.table}

\item{s}{vector}

\item{t}{vector}

\item{zp}{numeric}

\item{kf}{numeric}
}
\value{
data.table
}
\description{
Title
}
