% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildIMF.R
\name{buildIMF}
\alias{buildIMF}
\title{Title}
\usage{
buildIMF(
  s,
  t = NULL,
  dt = NULL,
  method = "emd",
  boundary = "wave",
  max.imf = 15,
  noise.type = "gaussian",
  noise.amp = 5e-08,
  trials = 10,
  stop.rule = "type5",
  plot = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{s}{numeric vector}

\item{t}{numeric vector}

\item{dt}{numeric}

\item{method}{string}

\item{boundary}{string}

\item{max.imf}{integer}

\item{noise.type}{string}

\item{noise.amp}{numeric}

\item{trials}{integer}

\item{stop.rule}{string}

\item{plot}{boolean}

\item{verbose}{boolean}
}
\value{
list
}
\description{
Title
}
