% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTS.R
\name{buildTS}
\alias{buildTS}
\title{buildTS}
\usage{
buildTS(
  x,
  dt,
  UN,
  Modes = NULL,
  Fmax = 25,
  Resample = TRUE,
  FlatZeros = TRUE,
  TrimZeros = TRUE,
  DetrendAT = TRUE,
  DetrendVT = TRUE,
  DetrendDT = TRUE,
  PadZeros = TRUE,
  TargetUnits = "mm",
  NW = 2048,
  OVLP = 75
)
}
\arguments{
\item{x}{data.table}

\item{dt}{numeric}

\item{UN}{character}

\item{Modes}{numeric}

\item{Fmax}{integer}

\item{Resample}{boolean}

\item{FlatZeros}{boolean}

\item{TrimZeros}{boolean}

\item{DetrendAT}{boolean}

\item{DetrendVT}{boolean}

\item{DetrendDT}{boolean}

\item{PadZeros}{boolean}

\item{TargetUnits}{character Units}

\item{NW}{integer Windows Length}

\item{OVLP}{integer}
}
\value{
data.table
}
\description{
buildTS
}
