% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Intensity.R
\name{get_Intensity}
\alias{get_Intensity}
\title{Title}
\usage{
get_Intensity(TSL, TargetUnits = "mm")
}
\arguments{
\item{TSL}{data.table}

\item{TargetUnits}{character}
}
\value{
list
}
\description{
Title
}
