% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Kmax.R
\name{get_kmax}
\alias{get_kmax}
\title{Title}
\usage{
get_kmax(.SD, TargetUnits = "mm", interval = FALSE)
}
\arguments{
\item{.SD}{data.table}

\item{TargetUnits}{character}

\item{interval}{boolean}
}
\description{
Title
}
