% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nd.R
\name{get_nd}
\alias{get_nd}
\title{Title}
\usage{
get_nd(.x = NULL, s = NULL, ts = NULL, Units = "mm", kh)
}
\arguments{
\item{.x}{data.table}

\item{s}{vector}

\item{ts}{vector}

\item{Units}{character}

\item{kh}{numeric}
}
\description{
Title
}
