% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Newmark.R
\name{get_Newmark}
\alias{get_Newmark}
\title{Title}
\usage{
get_Newmark(AT, t = NULL, dt = NULL, kh, TOL = 0.001, g = 9806.65, FULL = TRUE)
}
\arguments{
\item{AT}{data.table}

\item{t}{vector}

\item{dt}{double}

\item{kh}{double}

\item{TOL}{double}

\item{g}{double}

\item{FULL}{logical}
}
\description{
Title
}
