% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIM.R
\name{getIM}
\alias{getIM}
\title{Title}
\usage{
getIM(a = NULL, v = NULL, d = NULL, dt = NULL, UN = NULL, TargetUnits = "mm")
}
\arguments{
\item{a}{data.table Time Series of acceleration records}

\item{v}{data.table Time Series of velocity records}

\item{d}{data.table Time Series of displacement records}

\item{dt}{numeric Time step}

\item{UN}{character Units of source records}

\item{TargetUnits}{character Units of target records}
}
\value{
data.table
}
\description{
Title
}
