% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVDA.R
\name{getVDA}
\alias{getVDA}
\title{Title}
\usage{
getVDA(
  a,
  dt,
  UN,
  Amin = 0,
  Amax = Inf,
  TFT = NULL,
  DownFs = 0,
  UpFs = 0,
  DetrendAT = FALSE,
  DetrendVT = FALSE,
  DetrendDT = FALSE,
  Fpass_LP = 0,
  Fstop_LP = 0,
  Fpass_HP = 0,
  Fstop_HP = 0,
  RestoreScale = FALSE,
  TargetUnits = "mm",
  NW = 2048,
  OVLP = 75
)
}
\arguments{
\item{a}{data.table Time Series}

\item{dt}{numeric Time Step}

\item{UN}{character Units}

\item{Amin}{numeric Minimum Amplitude}

\item{Amax}{numeric Maximum Amplitude}

\item{TFT}{data.table Transfer Functions Table}

\item{DownFs}{integer Downsample Frequency}

\item{UpFs}{integer Upsample Frequency}

\item{DetrendAT}{boolean Detrend Acceleration Time Series}

\item{DetrendVT}{boolean Detrend Velocity Time Series}

\item{DetrendDT}{boolean Detrend Displacement Time Series}

\item{Fpass_LP}{numeric Low Pass Frequency}

\item{Fstop_LP}{numeric Low Stop Frequency}

\item{Fpass_HP}{numeric High Pass Frequency}

\item{Fstop_HP}{numeric High Stop Frequency}

\item{RestoreScale}{boolean}

\item{TargetUnits}{character Units}

\item{NW}{integer Windows Length}

\item{OVLP}{integer Overlap}
}
\value{
data.table
}
\description{
Title
}
