% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_IMF.R
\name{remove_IMF}
\alias{remove_IMF}
\title{Title}
\usage{
remove_IMF(X, EMD, removeIMF1 = 0, removeIMFn = 0)
}
\arguments{
\item{X}{vector. Time Series}

\item{EMD}{vector. Empirical Mode Decomposition of X}

\item{removeIMF1}{integer. Number of IMFs to remove from the beginning}

\item{removeIMFn}{integer. Number of IMFs to remove from the end}
}
\value{
list
}
\description{
Title
}
