% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_TS.R
\name{build_TS}
\alias{build_TS}
\title{buildTS}
\usage{
build_TS(
  x,
  ts = NULL,
  dt = NULL,
  Units,
  DetrendAT = FALSE,
  DetrendVT = FALSE,
  DetrendDT = FALSE,
  Fmax = 16,
  kNyq = 3.125,
  Resample = TRUE,
  LowPass = TRUE,
  TargetUnits = "mm",
  NW = 1024,
  OVLP = 75,
  Astop.AT = 1e-04,
  Apass.AT = 0.001,
  PreserveMaximum = FALSE,
  Output = NULL
)
}
\arguments{
\item{x}{data.table}

\item{ts}{numeric}

\item{dt}{numeric}

\item{Units}{character}

\item{DetrendAT}{boolean}

\item{DetrendVT}{boolean}

\item{DetrendDT}{boolean}

\item{Fmax}{integer}

\item{kNyq}{numeric}

\item{Resample}{boolean}

\item{LowPass}{boolean}

\item{TargetUnits}{character Units}

\item{NW}{integer Windows Length}

\item{OVLP}{integer}

\item{Astop.AT}{numeric}

\item{Apass.AT}{numeric}

\item{PreserveMaximum}{boolean}

\item{Output}{character}
}
\value{
list
}
\description{
buildTS
}
