% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_TS.R
\name{build_TS}
\alias{build_TS}
\title{buildTS}
\usage{
build_TS(
  x,
  ts = NULL,
  dt = NULL,
  Units,
  DetrendAT = FALSE,
  DetrendVT = FALSE,
  DetrendDT = FALSE,
  Fmax = 16,
  kNyq = 3.125,
  Resample = TRUE,
  TargetUnits = "mm",
  NW = 128,
  OVLP = 75,
  FlatZeros = FALSE,
  AstopAT = 1e-04,
  ApassAT = 0.001,
  TrimZeros = FALSE,
  Normalize = FALSE,
  Output = NULL
)
}
\arguments{
\item{x}{data.table}

\item{ts}{numeric}

\item{dt}{numeric}

\item{Units}{character}

\item{DetrendAT}{boolean}

\item{DetrendVT}{boolean}

\item{DetrendDT}{boolean}

\item{Fmax}{integer}

\item{kNyq}{numeric}

\item{Resample}{boolean}

\item{TargetUnits}{character Units}

\item{NW}{integer Windows Length}

\item{OVLP}{integer}

\item{FlatZeros}{boolean}

\item{AstopAT}{numeric}

\item{ApassAT}{numeric}

\item{TrimZeros}{boolean}

\item{Normalize}{boolean}

\item{Output}{character}
}
\value{
list
}
\description{
buildTS
}
