% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTS.R
\name{buildTS}
\alias{buildTS}
\title{buildTS}
\usage{
buildTS(
  x,
  ts = NULL,
  dt = NULL,
  Units,
  OrderTS = 2,
  OrderEMD = 2,
  Fmax = 16,
  Resample = TRUE,
  LowPass = TRUE,
  removeIMF1 = 0,
  removeIMFn = 0,
  TargetUnits = "mm",
  NW = 1024,
  OVLP = 75,
  Astop.AT = 1e-04,
  Apass.AT = 0.001
)
}
\arguments{
\item{x}{data.table}

\item{ts}{numeric}

\item{dt}{numeric}

\item{Units}{character}

\item{OrderTS}{integer}

\item{OrderEMD}{integer}

\item{Fmax}{integer}

\item{Resample}{boolean}

\item{LowPass}{boolean}

\item{removeIMF1}{boolean}

\item{removeIMFn}{boolean}

\item{TargetUnits}{character Units}

\item{NW}{integer Windows Length}

\item{OVLP}{integer}

\item{Astop.AT}{numeric}

\item{Apass.AT}{numeric}
}
\value{
list
}
\description{
buildTS
}
