% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTS.R
\name{buildTS}
\alias{buildTS}
\title{buildTS}
\usage{
buildTS(
  x,
  ts = NULL,
  dt = NULL,
  UN,
  Order = 2,
  Fmax = 16,
  Resample = TRUE,
  LowPass = FALSE,
  removeIMF1 = 0,
  removeIMFn = 0,
  TargetUnits = "mm",
  NW = 1024,
  OVLP = 75
)
}
\arguments{
\item{x}{data.table}

\item{ts}{numeric}

\item{dt}{numeric}

\item{UN}{character}

\item{Order}{integer}

\item{Fmax}{integer}

\item{Resample}{boolean}

\item{LowPass}{boolean}

\item{removeIMF1}{boolean}

\item{removeIMFn}{boolean}

\item{TargetUnits}{character Units}

\item{NW}{integer Windows Length}

\item{OVLP}{integer}
}
\value{
list
}
\description{
buildTS
}
