% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTS.R
\name{buildTS}
\alias{buildTS}
\title{Title}
\usage{
buildTS(
  a,
  dt,
  UN = NULL,
  hl = 500,
  DownFs = 0,
  UpFs = 0,
  FlatZerosAT = TRUE,
  DerivateDT = FALSE,
  DerivateVT = FALSE,
  DetrendAT = TRUE,
  DetrendVT = FALSE,
  DetrendDT = FALSE,
  Fpass_LP = 0,
  Fstop_LP = 0,
  Fpass_HP = 0,
  Fstop_HP = 0,
  TargetUnits = "mm",
  NW = 2048,
  OVLP = 75
)
}
\arguments{
\item{a}{data.table Time Series}

\item{dt}{numeric Time Step}

\item{UN}{character Units}

\item{hl}{numeric 1/Tolerance}

\item{DownFs}{integer Downsample Frequency}

\item{UpFs}{integer Upsample Frequency}

\item{FlatZerosAT}{boolean Flat Zeros Acceleration Time Series}

\item{DerivateDT}{boolean Derivate Displacements Time Series}

\item{DerivateVT}{boolean Derivate Velocity Time Series}

\item{DetrendAT}{boolean Detrend Acceleration Time Series}

\item{DetrendVT}{boolean Detrend Velocity Time Series}

\item{DetrendDT}{boolean Detrend Displacement Time Series}

\item{Fpass_LP}{numeric Low Pass Frequency}

\item{Fstop_LP}{numeric Low Stop Frequency}

\item{Fpass_HP}{numeric High Pass Frequency}

\item{Fstop_HP}{numeric High Stop Frequency}

\item{TargetUnits}{character Units}

\item{NW}{integer Windows Length}

\item{OVLP}{integer Overlap}
}
\value{
data.table
}
\description{
Title
}
