% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTS.R
\name{buildTS}
\alias{buildTS}
\title{buildTS}
\usage{
buildTS(
  x,
  dt,
  UN,
  Fmax = 25,
  Resample = TRUE,
  FlatZeros = TRUE,
  TrimZeros = TRUE,
  Detrend_AT = TRUE,
  Detrend_VT = FALSE,
  Detrend_DT = FALSE,
  PadZeros = TRUE,
  RemoveFirstIMF_AT = FALSE,
  RemoveLastIMF_AT = TRUE,
  RemoveFirstIMF_VT = FALSE,
  RemoveLastIMF_VT = TRUE,
  RemoveFirstIMF_DT = FALSE,
  RemoveLastIMF_DT = TRUE,
  Rebuild_AT = FALSE,
  TargetUnits = "mm",
  taper = 2,
  NW = 2048,
  OVLP = 75
)
}
\arguments{
\item{x}{data.table}

\item{dt}{numeric}

\item{UN}{character}

\item{Fmax}{integer}

\item{Resample}{boolean}

\item{FlatZeros}{boolean}

\item{TrimZeros}{boolean}

\item{Detrend_AT}{boolean}

\item{Detrend_VT}{boolean}

\item{Detrend_DT}{boolean}

\item{PadZeros}{boolean}

\item{RemoveFirstIMF_AT}{boolean}

\item{RemoveLastIMF_AT}{boolean}

\item{RemoveFirstIMF_VT}{boolean}

\item{RemoveLastIMF_VT}{boolean}

\item{RemoveFirstIMF_DT}{boolean}

\item{RemoveLastIMF_DT}{boolean}

\item{Rebuild_AT}{boolean}

\item{TargetUnits}{character Units}

\item{taper}{integer}

\item{NW}{integer Windows Length}

\item{OVLP}{integer}
}
\value{
data.table
}
\description{
buildTS
}
