% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTS.R
\name{buildTS}
\alias{buildTS}
\title{Title}
\usage{
buildTS(
  a,
  dt,
  UN,
  Fmax = 25,
  Resample = TRUE,
  FlatZeros = TRUE,
  RebuildAT = FALSE,
  Detrend = TRUE,
  TrimZeros = TRUE,
  PadZeros = TRUE,
  TargetUnits = "mm",
  NW = 2048,
  OVLP = 75
)
}
\arguments{
\item{a}{data.table Time Series}

\item{dt}{numeric Time Step}

\item{UN}{character Units}

\item{Fmax}{integer Maximum usable frequency}

\item{Resample}{boolean Add Zeros to Acceleration Time Series}

\item{FlatZeros}{boolean Flat Zeros Acceleration Time Series}

\item{RebuildAT}{boolean Derivate Displacements and Velocity Time Series}

\item{Detrend}{boolean Detrend Acceleration Time Series}

\item{TrimZeros}{boolean Trim Zeros Acceleration Time Series}

\item{PadZeros}{boolean Add Zeros to Acceleration Time Series}

\item{TargetUnits}{character Units}

\item{NW}{integer Windows Length}

\item{OVLP}{integer Overlap}
}
\value{
data.table
}
\description{
Title
}
