% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imf.R
\name{get_imf}
\alias{get_imf}
\title{Title}
\usage{
get_imf(
  .x = NULL,
  s = NULL,
  ts = NULL,
  dt = NULL,
  method = "emd",
  boundary = "wave",
  max.imf = 15,
  noise.type = "gaussian",
  noise.amp = 5e-08,
  trials = 10,
  stop.rule = "type5",
  verbose = FALSE
)
}
\arguments{
\item{.x}{data.table}

\item{s}{numeric vector}

\item{ts}{numeric vector}

\item{dt}{numeric}

\item{method}{string}

\item{boundary}{string}

\item{max.imf}{integer}

\item{noise.type}{string}

\item{noise.amp}{numeric}

\item{trials}{integer}

\item{stop.rule}{string}

\item{verbose}{boolean}
}
\value{
data.table
}
\description{
Title
}
