% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Intensity.R
\name{get_Intensity}
\alias{get_Intensity}
\title{Title}
\usage{
get_Intensity(
  TSL,
  TargetUnits = "mm",
  kh_values = c(0.01, 0.02, 0.05, 0.1, 0.15, 0.2, 0.25, 0.3, 0.35, 0.4, 0.45, 0.5, 0.55,
    0.6, 0.65, 0.7, 0.75, 0.8, 0.85)
)
}
\arguments{
\item{TSL}{data.table}

\item{TargetUnits}{character}

\item{kh_values}{vector}
}
\description{
Title
}
