% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ND.R
\name{get_ND}
\alias{get_ND}
\title{get_ND}
\usage{
get_ND(AT, t = NULL, dt = NULL, kh, TOL = 0.001, g = 9806.65, full = TRUE)
}
\arguments{
\item{AT}{data.table}

\item{t}{numeric}

\item{dt}{numeric}

\item{kh}{numeric}

\item{TOL}{numeric}

\item{g}{numeric}

\item{full}{boolean}
}
\value{
list
}
\description{
get_ND
}
