% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Spectra.R
\name{get_Spectra}
\alias{get_Spectra}
\title{Title}
\usage{
get_Spectra(.x = NULL, s = NULL, t = NULL, dt = NULL, Tn = NULL, xi = 0.05)
}
\arguments{
\item{.x}{data.table}

\item{s}{vector}

\item{t}{vector}

\item{dt}{double}

\item{Tn}{vector}

\item{xi}{double}
}
\value{
list
}
\description{
Title
}
